apps/fastapi-service/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
astroid==3.3.8
autopep8==2.3.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
isort==6.0.0
Jinja2==3.1.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
platformdirs==4.3.6
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pylint==3.3.4
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
sniffio==1.3.1
starlette==0.45.3
tomlkit==0.13.2
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
